third_party/blink/web_tests/fast/forms/date-multiple-fields/date-multiple-fields-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for date input fields from renderer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Function arguments are (value, min, max, [keySequence]).
Month
PASS stepUp("2000-05-01", "2000-05-01", "2000-10-31") is "2000-06-01"
PASS stepDown("2000-05-01", "2000-05-01", "2000-10-31") is "2000-10-01"
PASS stepUp("2000-05-01", "2000-02-01", "2000-05-31") is "2000-02-01"
PASS stepDown("2000-05-01", "2000-02-01", "2000-05-31") is "2000-04-01"
PASS stepUp("2000-05-01", "2000-08-01", "2000-10-31") is "2000-08-01"
PASS stepDown("2000-05-01", "2000-08-01", "2000-10-31") is "2000-10-01"
PASS stepUp("2000-05-01", "2000-03-01", "2000-03-31") is "2000-03-01"
PASS stepDown("2000-05-01", "2000-03-01", "2000-03-31") is "2000-03-01"
PASS stepUp("2000-05-01", "2000-03-01", "2001-05-31") is "2000-06-01"
PASS stepDown("2000-05-01", "2000-05-01", "2001-08-31") is "2000-04-01"
Day
PASS stepUp("2000-05-10", "2000-05-10", "2000-05-20") is "2000-05-11"
PASS stepDown("2000-05-10", "2000-05-10", "2000-05-20") is "2000-05-20"
PASS stepUp("2000-05-10", "2000-05-05", "2000-05-10") is "2000-05-05"
PASS stepDown("2000-05-10", "2000-05-05", "2000-05-10") is "2000-05-09"
PASS stepUp("2000-05-10", "2000-05-20", "2000-05-30") is "2000-05-20"
PASS stepDown("2000-05-10", "2000-05-20", "2000-05-30") is "2000-05-30"
PASS stepUp("2000-05-10", "2000-05-05", "2000-05-05") is "2000-05-05"
PASS stepDown("2000-05-10", "2000-05-05", "2000-05-05") is "2000-05-05"
PASS stepUp("2000-05-10", "2000-05-05", "2001-05-10") is "2000-05-11"
PASS stepDown("2000-05-10", "2000-05-10", "2001-05-15") is "2000-05-09"
PASS stepUp("2000-05-10", "2000-05-05", "2000-06-10") is "2000-05-11"
PASS stepDown("2000-05-10", "2000-05-10", "2000-06-15") is "2000-05-09"
Year
PASS stepUp("2000-01-01", "2000-01-01", "2009-12-31") is "2001-01-01"
PASS stepDown("2000-01-01", "2000-01-01", "2009-12-31") is "2009-01-01"
PASS stepUp("2000-01-01", "1990-01-01", "2000-12-31") is "1990-01-01"
PASS stepDown("2000-01-01", "1990-01-01", "2000-12-31") is "1999-01-01"
PASS stepUp("275760-01-01", "1990-01-01", null) is "1990-01-01"
PASS stepDown("1990-01-01", "1990-01-01", null) is "275760-01-01"
PASS stepUp("2000-01-01", null, "2000-12-31") is "0001-01-01"
PASS stepDown("0001-01-01", null, "2000-12-31") is "2000-01-01"
PASS stepUp("1990-01-01", "2000-01-01", "2009-12-31") is "2000-01-01"
PASS stepDown("1990-01-01", "2000-01-01", "2009-12-31") is "2009-01-01"
PASS stepUp("2020-01-01", "2000-01-01", "2000-12-31") is "2000-01-01"
PASS stepDown("2020-01-01", "2000-01-01", "2000-12-31") is "2000-01-01"


PASS successfullyParsed is true

TEST COMPLETE
